refactor(agent): extract AgentRuntime to agent::runtime; ignore test-results
- Move AgentRuntime enum from store::agents to src/agent/runtime.rs so
  it is accessible to both the store layer and the server handlers without
  a cross-crate import cycle.
- Update all driver files and handlers to import from crate::agent::runtime.
- Add acp_adaptor_binary() to AgentRuntime for driver-selection logic.
- Add test-results/, ui/test-results/, and **/.last-run.json to .gitignore.
