+++ src/agent/drivers/claude_raw.rs
@@ -5,7 +5,7 @@ use super::{command_exists, run_command, Driver, ParsedEvent, SpawnContext};
 use crate::agent::config::AgentConfig;
 use crate::agent::drivers::prompt::{build_base_system_prompt, PromptOptions};
 use crate::agent::runtime_status::{RuntimeAuthStatus, RuntimeStatus};
-use crate::store::agents::AgentRuntime;
+use crate::agent::AgentRuntime;
 
 pub struct ClaudeRawDriver;
