+++ src/agent/drivers/opencode_raw.rs
@@ -4,7 +4,7 @@ use super::prompt::{build_base_system_prompt, PromptOptions};
 use super::{command_exists, run_command, Driver, ParsedEvent, SpawnContext};
 use crate::agent::config::AgentConfig;
 use crate::agent::runtime_status::{RuntimeAuthStatus, RuntimeStatus};
-use crate::store::agents::AgentRuntime;
+use crate::agent::AgentRuntime;
 
 pub struct OpencodeRawDriver;
