+++ src/agent/mod.rs
@@ -3,8 +3,10 @@ pub mod config;
 pub mod drivers;
 pub mod lifecycle;
 pub mod manager;
+pub mod runtime;
 pub mod runtime_status;
 pub mod workspace;
 
 pub use lifecycle::AgentLifecycle;
 pub(crate) use lifecycle::NoopAgentLifecycle;
+pub use runtime::AgentRuntime;
