+++ src/main.rs
@@ -4,7 +4,8 @@ use std::sync::Arc;
 use chorus::agent::manager::AgentManager;
 use chorus::bridge;
 use chorus::server::build_router_with_lifecycle;
-use chorus::store::agents::{AgentRuntime, AgentStatus};
+use chorus::agent::AgentRuntime;
+use chorus::store::agents::AgentStatus;
 use chorus::store::channels::ChannelType;
 use chorus::store::messages::SenderType;
 use chorus::store::Store;
